split: test_weave_api into trace and legacy components
